testing answers controller

+++ rails/lecture/awesome_answer/app/controllers/questions_controller.rb
@@ -47,7 +47,12 @@ class QuestionsController < ApplicationController
 
     def destroy
         question = Question.find params[:id]
-        question.destroy
-        redirect_to questions_path
+        
+        if can?(:delete, question) == false
+            redirect_to root_path, alert: "Unauthorized user!"
+        else
+            question.destroy
+            redirect_to questions_path
+        end
     end
 end
